pyproject.toml
[project]
name = "mltechassessment"
version = "0.1.0"
description = ""
authors = [
    {name = "yeivillamil26",email = "yvillamilfranco@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "openai (>=1.86.0,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "streamlit-chat (>=0.1.1,<0.2.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "langchain-openai (>=0.3.21,<0.4.0)",
    "opensearch-py (>=2.8.0,<3.0.0)",
    "markdown (>=3.8,<4.0)",
    "markdownify (>=1.1.0,<2.0.0)",
#    "python-dotenv (>=1.1.0,<2.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mlflow (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
